Removed ambiguities in BTMessageProtocol / BTInfoProtocol pair (no more use of AnyObject). Also fixed a bug regarding the async protocol usage of sending category data items (category number for handler improperly set). Probably should have used Operations again as in BTImportExport.swift, but GCD is easier to use now. Had to make sure any construction of WKWebViews happened on the main thread (this is a UI element, albeit hidden with zero frame). A background queue (global) was used - we should try concurrent queues for better performance. The UI still does not look right (have a chance to update smoothly). So there is still a bug or two lurking.
Also, in InfoItem VC, we are reminded that there remains a (LARGE) bug of date ranges crashing from extraInfo due to the design limits of the date range concept itself (consider Mar 28 - Apr 3; the month goes up (OK) but the day goes down (NOT OK for a CountableClosedRange) - we need a custom struct or enum to better deal with this. Check out the .now() time enums.